Create a dedicated fuzz package
