+++ fuzz/build.gradle
@@ -1,0 +1,9 @@
+dependencies {
+    implementation project(':bls')
+    implementation project(':ethereum:core')
+    implementation project(':ethereum:datastructures')
+    implementation project(':ssz')
+    implementation project(':util')
+
+    implementation 'org.apache.tuweni:tuweni-bytes'
+}
