+++ fuzz/src/main/java/tech/pegasys/teku/fuzz/FuzzUtil.java
@@ -11,35 +11,30 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.core;
+package tech.pegasys.teku.fuzz;
 
-import com.google.common.primitives.UnsignedLong;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
-import tech.pegasys.teku.bls.BLSSignature;
+import tech.pegasys.teku.core.BlockProcessorUtil;
+import tech.pegasys.teku.core.StateTransition;
+import tech.pegasys.teku.core.StateTransitionException;
 import tech.pegasys.teku.core.exceptions.BlockProcessingException;
-import tech.pegasys.teku.datastructures.blocks.BeaconBlock;
-import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.datastructures.operations.Attestation;
-import tech.pegasys.teku.datastructures.operations.AttesterSlashing;
-import tech.pegasys.teku.datastructures.operations.Deposit;
-import tech.pegasys.teku.datastructures.operations.ProposerSlashing;
-import tech.pegasys.teku.datastructures.operations.SignedVoluntaryExit;
-import tech.pegasys.teku.datastructures.operations.VoluntaryExit;
 import tech.pegasys.teku.datastructures.state.BeaconState;
-import tech.pegasys.teku.datastructures.state.BeaconStateImpl;
 import tech.pegasys.teku.datastructures.util.BeaconStateUtil;
 import tech.pegasys.teku.datastructures.util.CommitteeUtil;
 import tech.pegasys.teku.datastructures.util.SimpleOffsetSerializer;
-import tech.pegasys.teku.ssz.SSZTypes.SSZContainer;
+import tech.pegasys.teku.fuzz.input.AttestationFuzzInput;
+import tech.pegasys.teku.fuzz.input.AttesterSlashingFuzzInput;
+import tech.pegasys.teku.fuzz.input.BlockFuzzInput;
+import tech.pegasys.teku.fuzz.input.BlockHeaderFuzzInput;
+import tech.pegasys.teku.fuzz.input.DepositFuzzInput;
+import tech.pegasys.teku.fuzz.input.ProposerSlashingFuzzInput;
+import tech.pegasys.teku.fuzz.input.VoluntaryExitFuzzInput;
 import tech.pegasys.teku.ssz.SSZTypes.SSZList;
 import tech.pegasys.teku.ssz.sos.ReflectionInformation;
-import tech.pegasys.teku.ssz.sos.SimpleOffsetSerializable;
 import tech.pegasys.teku.util.config.Constants;
 
 // TODO a Java FuzzHarness interface? - that way type safety can be checked at compile time
@@ -303,345 +298,6 @@ public class FuzzUtil {
     } catch (BlockProcessingException e) {
       // "expected error"
       return Optional.empty();
-    }
-  }
-
-  /** ******************** Input Classes ********************* */
-
-  // TODO common abstract class for all operations that are state + op?
-  // TODO move to separate package?
-  // NOTE: not obvious how to have a generic "OperationFuzzInput" class because the get_fixed_parts
-  // and get_variable_parts
-  // implementations can be different
-  public static class AttestationFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    // TODO should this be a BeaconState or BeaconStateImpl?
-    private BeaconStateImpl state;
-    private Attestation attestation;
-
-    public AttestationFuzzInput(final BeaconStateImpl state, final Attestation attestation) {
-      this.state = state;
-      this.attestation = attestation;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    public AttestationFuzzInput() {
-      this(new BeaconStateImpl(), new Attestation());
-    }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + attestation.getSSZFieldCount();
-    }
-
-    // Since its both fields are variable we leave untouched?
-    /*@Override
-    public List<Bytes> get_fixed_parts() {
-    List<Bytes> fixedPartsList = new ArrayList<>();
-    fixedPartsList.addAll(state.get_fixed_parts());
-    fixedPartsList.addAll(attestation.get_fixed_parts());
-    return fixedPartsList;
-    }*/
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      // Because we know both fields are variable and registered, we can just serialize.
-      return List.of(
-          SimpleOffsetSerializer.serialize(state), SimpleOffsetSerializer.serialize(attestation));
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public Attestation getAttestation() {
-      return attestation;
-    }
-
-    public BeaconState getState() {
-      return state;
-    }
-  }
-
-  public static class AttesterSlashingFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    // TODO should this be a BeaconState or BeaconStateImpl?
-    private BeaconStateImpl state;
-    private AttesterSlashing attester_slashing;
-
-    public AttesterSlashingFuzzInput(
-        final BeaconStateImpl state, final AttesterSlashing attester_slashing) {
-      this.state = state;
-      this.attester_slashing = attester_slashing;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    // public AttesterSlashingFuzzInput() {
-    //  this(new BeaconStateImpl(), new AttesterSlashing());
-    // }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + attester_slashing.getSSZFieldCount();
-    }
-
-    // Since its both fields are variable we leave untouched?
-    /*@Override
-    public List<Bytes> get_fixed_parts() {
-    List<Bytes> fixedPartsList = new ArrayList<>();
-    fixedPartsList.addAll(state.get_fixed_parts());
-    fixedPartsList.addAll(attester_slashing.get_fixed_parts());
-    return fixedPartsList;
-    }*/
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      // Because we know both fields are variable and registered, we can just serialize.
-      return List.of(
-          SimpleOffsetSerializer.serialize(state),
-          SimpleOffsetSerializer.serialize(attester_slashing));
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public AttesterSlashing getAttester_slashing() {
-      return attester_slashing;
-    }
-
-    public BeaconState getState() {
-      return state;
-    }
-  }
-
-  public static class BlockFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    private BeaconStateImpl state;
-    private SignedBeaconBlock signed_block;
-
-    public BlockFuzzInput(final BeaconStateImpl state, final SignedBeaconBlock signed_block) {
-      this.state = state;
-      this.signed_block = signed_block;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    public BlockFuzzInput() {
-      this(new BeaconStateImpl(), new SignedBeaconBlock(new BeaconBlock(), BLSSignature.empty()));
-    }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + signed_block.getSSZFieldCount();
-    }
-
-    // Since its both fields are variable we leave untouched?
-    /*@Override
-    public List<Bytes> get_fixed_parts() {
-    List<Bytes> fixedPartsList = new ArrayList<>();
-    fixedPartsList.addAll(state.get_fixed_parts());
-    fixedPartsList.addAll(signed_block.get_fixed_parts());
-    return fixedPartsList;
-    }*/
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      // Because we know both fields are variable and registered, we can just serialize.
-      return List.of(
-          SimpleOffsetSerializer.serialize(state), SimpleOffsetSerializer.serialize(signed_block));
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public SignedBeaconBlock getSigned_block() {
-      return signed_block;
-    }
-
-    public BeaconState getState() {
-      return state;
-    }
-  }
-
-  /**
-   * Note: BlockHeader fuzzing target accepts a block as input (not a SignedBeaconBlock or
-   * BeaconBlockHeader)
-   */
-  public static class BlockHeaderFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    // TODO should this be a BeaconState or BeaconStateImpl?
-    private BeaconStateImpl state;
-    private BeaconBlock block;
-
-    public BlockHeaderFuzzInput(final BeaconStateImpl state, final BeaconBlock block) {
-      this.state = state;
-      this.block = block;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    public BlockHeaderFuzzInput() {
-      this(new BeaconStateImpl(), new BeaconBlock());
-    }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + block.getSSZFieldCount();
-    }
-
-    // Since its both fields are variable we leave untouched?
-    /*@Override
-    public List<Bytes> get_fixed_parts() {
-    List<Bytes> fixedPartsList = new ArrayList<>();
-    fixedPartsList.addAll(state.get_fixed_parts());
-    fixedPartsList.addAll(block.get_fixed_parts());
-    return fixedPartsList;
-    }*/
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      // Because we know both fields are variable and registered, we can just serialize.
-      return List.of(
-          SimpleOffsetSerializer.serialize(state), SimpleOffsetSerializer.serialize(block));
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public BeaconBlock getBlock() {
-      return block;
-    }
-
-    public BeaconState getState() {
-      return state;
-    }
-  }
-
-  public static class DepositFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    // TODO should this be a BeaconState or BeaconStateImpl?
-    private BeaconStateImpl state;
-    private Deposit deposit;
-
-    public DepositFuzzInput(final BeaconStateImpl state, final Deposit deposit) {
-      this.state = state;
-      this.deposit = deposit;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    public DepositFuzzInput() {
-      this(new BeaconStateImpl(), new Deposit());
-    }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + deposit.getSSZFieldCount();
-    }
-
-    @Override
-    public List<Bytes> get_fixed_parts() {
-      List<Bytes> fixedPartsList = new ArrayList<>();
-      fixedPartsList.add(Bytes.EMPTY);
-      fixedPartsList.add(SimpleOffsetSerializer.serialize(deposit));
-      return fixedPartsList;
-    }
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      return List.of(SimpleOffsetSerializer.serialize(state), Bytes.EMPTY);
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public Deposit getDeposit() {
-      return deposit;
-    }
-
-    public BeaconState getState() {
-      return state;
-    }
-  }
-
-  public static class ProposerSlashingFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    // TODO should this be a BeaconState or BeaconStateImpl?
-    private BeaconStateImpl state;
-    private ProposerSlashing proposer_slashing;
-
-    public ProposerSlashingFuzzInput(
-        final BeaconStateImpl state, final ProposerSlashing proposer_slashing) {
-      this.state = state;
-      this.proposer_slashing = proposer_slashing;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    // public ProposerSlashingFuzzInput() {
-    //  this(new BeaconStateImpl(), new ProposerSlashing());
-    // }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + proposer_slashing.getSSZFieldCount();
-    }
-
-    @Override
-    public List<Bytes> get_fixed_parts() {
-      List<Bytes> fixedPartsList = new ArrayList<>();
-      fixedPartsList.add(Bytes.EMPTY);
-      fixedPartsList.add(SimpleOffsetSerializer.serialize(proposer_slashing));
-      return fixedPartsList;
-    }
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      return List.of(SimpleOffsetSerializer.serialize(state), Bytes.EMPTY);
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public ProposerSlashing getProposer_slashing() {
-      return proposer_slashing;
-    }
-
-    public BeaconState getState() {
-      return state;
-    }
-  }
-
-  public static class VoluntaryExitFuzzInput implements SimpleOffsetSerializable, SSZContainer {
-
-    // TODO should this be a BeaconState or BeaconStateImpl?
-    private BeaconStateImpl state;
-    private SignedVoluntaryExit exit;
-
-    public VoluntaryExitFuzzInput(final BeaconStateImpl state, final SignedVoluntaryExit exit) {
-      this.state = state;
-      this.exit = exit;
-    }
-
-    // NOTE: empty constructor is needed for reflection/introspection
-    // TODO how is new ReflectionInformation(VoluntaryExit.class) supposed to work when
-    // it cals newInstance() with no arguments but VoluntaryExit has no 0 arg constructor?
-    public VoluntaryExitFuzzInput() {
-      this(
-          new BeaconStateImpl(),
-          new SignedVoluntaryExit(
-              new VoluntaryExit(UnsignedLong.valueOf(0), UnsignedLong.valueOf(0)),
-              BLSSignature.empty()));
-    }
-
-    @Override
-    public int getSSZFieldCount() {
-      return state.getSSZFieldCount() + exit.getSSZFieldCount();
-    }
-
-    @Override
-    public List<Bytes> get_fixed_parts() {
-      List<Bytes> fixedPartsList = new ArrayList<>();
-      fixedPartsList.add(Bytes.EMPTY);
-      fixedPartsList.add(SimpleOffsetSerializer.serialize(exit));
-      return fixedPartsList;
-    }
-
-    @Override
-    public List<Bytes> get_variable_parts() {
-      return List.of(SimpleOffsetSerializer.serialize(state), Bytes.EMPTY);
-    }
-
-    /** ******************* * GETTERS & SETTERS * * ******************* */
-    public SignedVoluntaryExit getExit() {
-      return exit;
-    }
-
-    public BeaconState getState() {
-      return state;
     }
   }
 }
