+++ settings.gradle
@@ -10,6 +10,7 @@ include 'data:provider'
 include 'data:beaconrestapi'
 include 'data:recorder'
 include 'events'
+include 'fuzz'
 include 'logging'
 include 'networking:eth2'
 include 'networking:p2p'
